Update profile_tree_gadget for new API. Also a compile-time fix for CUDA 7.0 (no min/max overload for size_t args).

+++ tests/AABB_intersect/profile.cuh
@@ -71,7 +71,7 @@ float profile_gpu(const thrust::device_vector<Ray>& rays,
     cudaEventCreate(&start);
     cudaEventCreate(&stop);
 
-    const int num_blocks = min((rays.size() + 127) / 128, (size_t)48);
+    const int num_blocks = min((unsigned long long)((rays.size() + 127) / 128), (unsigned long long)48);
 
     cudaEventRecord(start);
     intersect_kernel<<<num_blocks, 128>>>(
